project/somaz-bigquery/cloud-functions/analytics-to-sheet-new-web-visitors/requirements.txt
google-api-python-client==2.33.0
gspread==5.3.2
google-auth==2.6.6
google-analytics-data==0.8.0
